+++ src/data/vcc_embedding_module.py
@@ -77,11 +77,13 @@ class VCCEmbeddingDataset(Dataset):
         #         {control_expression.shape[0], control_expression.shape[0], ko_gene_data.shape[0]}"
         #         )
 
-        self.control_expression = control_expression.to_numpy()
+        self.control_expression = control_expression.to_numpy().astype(np.float16)
         self.perturbed_genes = perturbed_genes
         self.gene_embeddings = gene_embeddings.to_pandas().set_index("gene_name")
         self.ko_expression = (
-            ko_expression.to_numpy() if (ko_expression is not None) else None
+            ko_expression.to_numpy().astype(np.float16)
+            if (ko_expression is not None)
+            else None
         )
 
         del control_expression, perturbed_genes, gene_embeddings
@@ -101,14 +103,14 @@ class VCCEmbeddingDataset(Dataset):
         Retrieves a single data sample
         """
 
-        exp_input = self.control_expression[index, :].astype(np.float16)
+        exp_input = self.control_expression[index, :]
         genes = self.perturbed_genes[index]
         gene_input = self.gene_embeddings.loc[genes].values.astype(np.float16)
 
         if self.stage == "predict":
             pred_input = []
         else:
-            pred_input = self.ko_expression[index, :].astype(np.float16)
+            pred_input = self.ko_expression[index, :]
 
         return {"ko_vec": gene_input, "exp_vec": exp_input}, pred_input
